backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pydantic-settings
aiosqlite
python-multipart
apscheduler
rapidfuzz
mail-parser
python-dateutil
stripe
PyJWT[crypto]
google-auth
google-auth-oauthlib
google-api-python-client
